+++ src/main/java/com/apartment/apart/domain/community/CommunityRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CommunityRepository extends JpaRepository<Community, Integer> {
+public interface CommunityRepository extends JpaRepository<Community, Long> {
     Page<Community> findAll(Pageable pageable);
     Page<Community> findAll(Specification<Community> spec, Pageable pageable);
 }
